refactor(redir): move resolver code to redirects
